requirements.txt
grpcio
randomproto
protobuf
ipykernel
toolz